backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
schedule==1.2.0
requests==2.31.0
numpy==1.26.1
pandas==2.1.2
scikit-learn==1.3.2
geopy==2.4.0
geopandas==0.14.0
matplotlib==3.8.1
folium==0.14.0
shapely==2.0.2
python-multipart==0.0.6
pydantic==2.4.2
pytest==7.4.3 